Empty input buffer
Apparently using fgets after Scanf yields unknown results. In my case, the first fgets used to get skipped in the add method.

That was probably because of a lingering \n in the input buffer.

I placed a getchar() after scanf in the switching method to resolve that issue.


I also changed the gets method to fgets. don't think that impacted the program in any way